+++ delta/plugins/blazegraph/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/blazegraph/model/BlazegraphViewSearchParams.scala
@@ -29,7 +29,7 @@ import monix.bio.UIO
 final case class BlazegraphViewSearchParams(
     project: Option[ProjectRef] = None,
     deprecated: Option[Boolean] = None,
-    rev: Option[Long] = None,
+    rev: Option[Int] = None,
     createdBy: Option[Subject] = None,
     updatedBy: Option[Subject] = None,
     types: Set[Iri] = Set.empty[Iri],
